gui: Moved all conversion of current limit from mA->A to nice_spin_box. Removed const_iterator as class member and uses it only in reimplemented stepBy function.

+++ gui/qt/main_window.h
@@ -217,7 +217,7 @@ public:
   void set_current_limit_forward_spinbox(uint16_t code);
   void set_current_limit_reverse_spinbox(uint16_t code);
   void recommended_codes();
-  QMultiMap<int, double> recommended_current_limit_codes;
+  QMultiMap<int, int> recommended_current_limit_codes;
 
 private:
   // Helper method for setting the index of a combo box, given the desired
